Update Statistic function
