+++ script/statistic/__init__.py
@@ -1,0 +1,44 @@
+from statistic.base import statistics_calculate
+from statistic.stat_correlation import stat_correlation
+from statistic.stat_standard_deviation import stat_standard_deviation
+from statistic.stat_z_score import stat_z_score
+from statistic.stat_Basic import stat_mean, stat_median, stat_max, stat_min, stat_sum
+from statistic.stat_covariance import stat_covariance
+from statistic.stat_autocorrelation import stat_autocorrelation
+from statistic.stat_diff import stat_diff
+from statistic.stat_resample import stat_resample
+from statistic.stat_rolling import stat_rolling
+from statistic.stat_functional_response import stat_functional_response
+from statistic.stat_hellinger_distance import stat_hellinger_distance
+from statistic.stat_three_cornered_hat import stat_three_cornered_hat
+from statistic.stat_partial_least_squares_regression import stat_partial_least_squares_regression
+from statistic.stat_mann_kendall_trend_test import stat_mann_kendall_trend_test
+from statistic.stat_False_Discovery_Rate import stat_False_Discovery_Rate
+from statistic.stat_anova import stat_anova
+from statistic.stat_mann_kendall_trend_test import stat_mann_kendall_trend_test
+
+statistics_calculate.stat_correlation = stat_correlation
+statistics_calculate.stat_standard_deviation = stat_standard_deviation
+statistics_calculate.stat_z_score = stat_z_score
+statistics_calculate.stat_mean = stat_mean
+statistics_calculate.stat_median = stat_median
+statistics_calculate.stat_max = stat_max
+statistics_calculate.stat_min = stat_min
+statistics_calculate.stat_sum = stat_sum
+statistics_calculate.stat_covariance = stat_covariance
+statistics_calculate.stat_autocorrelation = stat_autocorrelation
+statistics_calculate.stat_diff = stat_diff
+statistics_calculate.stat_resample = stat_resample
+statistics_calculate.stat_rolling = stat_rolling
+statistics_calculate.stat_functional_response = stat_functional_response
+statistics_calculate.stat_hellinger_distance = stat_hellinger_distance
+statistics_calculate.stat_three_cornered_hat = stat_three_cornered_hat
+statistics_calculate.stat_partial_least_squares_regression = stat_partial_least_squares_regression
+statistics_calculate.stat_mann_kendall_trend_test = stat_mann_kendall_trend_test
+statistics_calculate.stat_False_Discovery_Rate = stat_False_Discovery_Rate
+statistics_calculate.stat_anova = stat_anova
+statistics_calculate.stat_mann_kendall_trend_test = stat_mann_kendall_trend_test
+
+
+# 导出所有内容
+__all__ = ['statistics_calculate']
